+++ app/routers/horario.py
@@ -18,10 +18,10 @@ router = APIRouter(
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=dict)
 async def crear_horario(
     horario: HorarioImport,
-    #token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    #if not validate_role(token): 
-    # raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="UNAUTHORIZED")
+    if not validate_role(token, ["admin"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
 
     horario_id = insert_horario(horario)
 
@@ -37,10 +37,10 @@ async def crear_horario(
 
 @router.get("/", response_model=List[HorarioOut], status_code=status.HTTP_200_OK)
 async def ver_horarios(
-    # token: str = Depends(oauth2_scheme)
-    ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="UNAUTHORIZED")
+    token: str = Depends(oauth2_scheme)
+):
+    if not validate_role(token, ["admin", "directivo", "profesor"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
 
     return read_all_horarios()
 
@@ -49,10 +49,10 @@ async def ver_horarios(
 async def actualizar_horario(
     id:int,
     horario: HorarioImport,
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=401, detail="UNAUTHORIZED")
+    if not validate_role(token, ["admin"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
 
     updated = update_horario(id, horario)
 
@@ -68,10 +68,10 @@ async def actualizar_horario(
 @router.delete("/{id}/", status_code=status.HTTP_200_OK)
 async def borrar_horario(
     id: int, 
-    #token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
     ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="UNAUTHORIZED")
+    if not validate_role(token, ["admin"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
 
     deleted = delete_horario(id)
     if not deleted:
