+++ app/routers/previ.py
@@ -15,12 +15,11 @@ async def crear_previ(
     id_directivo: int,
     id_expediente: int,
     previ : PreviImport,
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
+    if not validate_role(token, ["admin", "directivo"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     
-    # Aquí iría la lógica para obtener los datos de "previ" desde la base de datos
     previ_id = insert_previ(id_directivo, id_expediente, previ)
     if previ_id == -1:
         raise HTTPException(
@@ -34,36 +33,29 @@ async def crear_previ(
 @router.get("/expediente/{id_expediente}", response_model=List[PreviOut], status_code=status.HTTP_200_OK)
 async def ver_previes_por_expediente(
     id_expediente: int, 
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
-    
-    # Aquí iría la lógica para obtener los datos de "previ" desde la base de datos
+    if not validate_role(token, ["admin", "directivo"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     return read_previ_by_expediente(id_expediente)
 
 
 @router.get("/directivo/{id_directivo}", response_model=List[PreviOut], status_code=status.HTTP_200_OK)
 async def ver_previes_por_directivo(
     id_directivo: int, 
-    #TODO solo funciona al pasarle id expediente NOSE POR QUE
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
-    
-    # Aquí iría la lógica para obtener los datos de "previ" desde la base de datos
+    if not validate_role(token, ["admin", "directivo"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     return read_previ_by_directivo(id_directivo)
 
 
 @router.get("/", response_model=List[PreviOut], status_code=status.HTTP_200_OK)
 async def ver_todos_los_previes(
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
-    
-    # Aquí iría la lógica para obtener los datos de "previ" desde la base de datos
+    if not validate_role(token, ["admin", "directivo"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     return read_all_previes()
 
 
@@ -71,12 +63,10 @@ async def ver_todos_los_previes(
 async def actualizar_previ(
     id_previ: int, 
     previ: PreviImport, 
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
-    
-    # Aquí iría la lógica para actualizar los datos de "previ" en la base de datos
+    if not validate_role(token, ["admin", "directivo"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     updated = update_previ(id_previ, previ)
     if not updated:
         raise HTTPException(
@@ -89,12 +79,10 @@ async def actualizar_previ(
 @router.delete("/{id_previ}", status_code=status.HTTP_200_OK, response_model=dict)
 async def borrar_previ(
     id_previ: int, 
-    # token: str = Depends(oauth2_scheme)
+    token: str = Depends(oauth2_scheme)
 ):
-    # if not validate_role(token):
-    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No tienes permisos para acceder a esta ruta.")
-    
-    # Aquí iría la lógica para eliminar el "previ" de la base de datos
+    if not validate_role(token, ["admin"]):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Sin permisos")
     deleted = delete_previ(id_previ)
     if not deleted:
         raise HTTPException(
